refactor package, rename app.py to __init__.py
