+++ .gitignore
@@ -1,5 +1,5 @@
 # credentials
-flaskApp/firebase/serviceAccountKey.json
+hua/firebase/serviceAccountKey.json
 email_credentials.py
 .env
