+++ hua/static/dashboard/dashboard.js
