+++ hua/templates/home.html
@@ -1,0 +1,80 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Home Page</title>
+</head>
+<body>
+    <h1>Welcome, User!</h1>
+    <h3>To access your dashboard and be able to upload audio files, you need to input to your institution's access code</h3>
+    <div class="modal-button" id="openModal">Click to Enter Access Code</div>
+    <div id="access-code-modal" class="modal">
+        <div class="modal-content">
+            <span class="close">&times;</span>
+            <h3>Enter Access Code</h3>
+            <input type="text" id="accessCode" placeholder="Enter Access Code">
+            <button class="submit-btn" id="submitCode">Submit</button>
+        </div>
+    </div>
+    <div>
+        <div>
+            <a href="{{ url_for('pi_access_request') }}">Request PI Access</a>
+        </div>
+    </div>
+</body>
+</html>
+
+<script>
+var modal = document.getElementById("access-code-modal");
+var btn = document.getElementById("openModal");
+var span = document.getElementsByClassName("close")[0];
+btn.onclick = function() {
+    modal.style.display = "block";
+}
+span.onclick = function() {
+    modal.style.display = "none";
+}
+window.onclick = function(event) {
+    if (event.target == modal) {
+        modal.style.display = "none";
+    }
+}
+</script>
+
+<style>
+    .modal-button{
+        border: 2px dashed #ccc;
+        padding: 20px;
+        width: 250px;
+        height: 200px;
+        text-align: center;
+        border-radius: 10px;
+        font-size: 18px;
+        margin-top: 100px;
+        cursor: pointer;
+    }
+    .modal-button:hover{
+        background-color: #f4f4f9;
+    }
+
+    .modal {
+            display: none;
+            position: fixed;
+            z-index: 1;
+            left: 0;
+            top: 0;
+            width: 100%;
+            height: 100%;
+            background-color: rgba(0,0,0,0.5);}
+
+    .modal-content {
+        background-color: white;
+        margin: 15% auto;
+        padding: 20px;
+        border: 1px solid #888;
+        width: 300px;
+        border-radius: 10px;
+    }
+
+</style>
